Move globbing-related functions to engine/system.h
